Do some code reorganization per Fernando's comments
